Ai model/Faces Detectors/webapp/requirements.txt
flask
opencv-python
numpy
matplotlib
scikit-learn
joblib
fer
tensorflow-cpu
Pillow
requests
